Put Contact list item into its own component.
