+++ src/components/ContactList/ContactList.module.css
@@ -5,7 +5,7 @@
 
 .item {
   display: flex;
-  justify-content: space-between;
+  /* justify-content: space-between; */
   align-items: center;
   margin-bottom: 15px;
   padding: 10px 10px;
@@ -14,7 +14,7 @@
   background-color: #ffffff;
   box-shadow: -1px 2px 10px 2px rgba(34, 60, 80, 0.2);
 }
-
+/* 
 .item p {
   margin: 0 0 10px;
 }
@@ -37,7 +37,7 @@
   color: #ffffff;
 
   background-color: #ff8a80;
-}
+} */
 
 .notification {
   list-style-type: none;
